refactor exceptions
